Shift utiltiy functions to utils file
