+++ frontend/src/lib/components/SaveFunctionModal.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
     import { onMount, onDestroy } from 'svelte';
     import type * as Monaco from 'monaco-editor/esm/vs/editor/editor.api';
-    import { getApiUrl, fetchApi } from "$lib/utils/api";
+    import { fetchApi } from "$lib/utils/api";
     
     export let show = false;
     export let onClose: () => void;
